requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
stripe==5.5.0
passlib==1.7.4
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
click==8.1.6
MarkupSafe==2.1.3
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
click==8.1.6
MarkupSafe==2.1.3
